Unit test refactor for login and signup
